backend/requirements.txt
# ============================================================
# requirements.txt - AgenticLearn Backend Dependencies
# Python 3.10+ | Optimized for RTX 3050, 8GB RAM
# ============================================================

# FastAPI & Server
fastapi==0.110.0
uvicorn[standard]==0.27.1
python-multipart==0.0.9

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.28

# PDF Extraction
PyMuPDF==1.23.26

# ML / NLP
torch==2.2.1
transformers==4.38.2
datasets==2.18.0
scikit-learn==1.4.1.post1
numpy==1.26.4
sentencepiece==0.2.0
accelerate==0.27.2

# Gemini AI
google-generativeai==0.4.1

# Utilities
pydantic==2.6.3
python-dotenv==1.0.1
